Program for Profit and Loss
Here i committed a Profit and Loss Program in C language Repository. The logic of code is based on ladder if else. In that code i generalize all perspectives of profit and loss. So User enters both Cost price and Selling price he got "No Profit No Loss" as a Output.