Do while Example
STATEMENT: Write a program to input the code of a particular item, the quality purchased and the rate. Then ask if any more items are to be purchased. If not, then calculate the total purchase price and print it along with the gift to be presented. The gifts to the customer are given in the following manner.

100 and above but less than 500       -      A key ring   
500 and above but less than 1000     -       A leather purse
1000 and above                                 -       A Pocket Calculator

Before ending the program a "Thank You" statement should be printed to all the customers.