Table Program in C
A table (or multiplication table) of numbers is generated by multiplying a constant number with an iterative number from 1 to 10 to get the table. In other words, we can get the table of a number by multiplying the given number with counting from 1, 2, 3, 4, 5, ..., 9, 10. And on each iteration, the value of counting is incremented by 1, which goes up to 10, to print a complete table.