C program to print sum of series 1/2+1/4+1/6+…+1/n
For loop is used to compute the summation of each integer value. Initialize the value of ‘i’ variable to 1. Check the condition that the value of ‘i’ variable is less than or equal to the value of ‘number’ variable. If the condition is true, then execute the iteration of the loop and add the sum of series.

Nested if else condition statement is used to check the value of ‘i’ variable is equal to 1. If the condition is true then execute the statement. Otherwise, if the condition is false, execute the else if statement. Check the condition that the value of ‘i’ variable is equal to the value of ‘number’ variable.

If the condition is true then execute the statement and compute the sum of series. Otherwise, if the condition is false, then execute else statement. Print the sum of series using printf statement.